+++ NEWS.md
@@ -7,6 +7,11 @@
 
 ### Technical / Other changes
 -->
+## 2021-02-09 kf-portal-ui 2.34.0
+### Features
+- [2437](https://github.com/kids-first/kf-portal-ui/issues/2437) Feature (Ontology browser): Add Any Of / All Of / Not in the modal
+- [2913](https://github.com/kids-first/kf-portal-ui/issues/2913) Feature: Add banner to advertise KF Forum
+
 ## 2021-01-21 kf-portal-ui 2.33.0
 ### Technical / Other changes
 - [2914](https://github.com/kids-first/kf-portal-ui/issues/2914) Fix (VariantDB):  Update numbers.